Add protocol provider resource.
Add a resource that will return the current time, built on the new
terraform-plugin-go module. Because this module is private right now,
you'll need to set up Go to use it. The first step is to tell Go not to
use the GOSUMDB or GOPROXY for it:

go env -w GOPRIVATE=github.com/hashicorp/terraform-plugin-go

Next you'll still need to tell Go how to authenticate with GitHub, so it
can fetch the module from GitHub:

git config --global url."https://{username}:{github_personal_access_token}@github.com".insteadOf "https://github.com"

Now Go should Just Work with module, as usual.

This also renames some of the packages. The provider implementations are
now stored in packages specific to the SDK version the providers are
built with, and the backend database has been extracted into a new
backend package that all these provider implementations can rely on.