Add terraform-plugin-framework-timetypes smoke testing (#252)
The `timetypes` library is being extended with new Go `time.Duration` types, so now seems a good as a time as any to ensure there is some lightweight integration testing of terraform-plugin-framework-timetypes. That Go module could reach out to this testing similar to the Continuous Integration setup of terraform-plugin-go and other plugin Go modules.